move options to their own file
