+++ packages/insomnia-inso/src/__tests__/util.test.js
@@ -1,26 +1,5 @@
 // @flow
-import commander from 'commander';
-import { getAllOptions, exit, logErrorExit1, getDefaultAppDataDir } from '../util';
-
-describe('getAllOptions()', () => {
-  it('should combine options from all commands into one object', () => {
-    const command = new commander.Command('command');
-
-    command
-      .command('subCommand')
-      .option('-s, --subCmd')
-      .action(cmd => {
-        expect(getAllOptions(cmd)).toEqual({
-          global: true,
-          subCmd: true,
-        });
-      });
-
-    const parent = new commander.Command().option('-g, --global').addCommand(command);
-
-    parent.parse('node test command subCommand --global --subCmd'.split(' '));
-  });
-});
+import { exit, logErrorExit1, getDefaultAppDataDir } from '../util';
 
 describe('exit()', () => {
   it('should exit 0 if successful result', async () => {
