+++ packages/insomnia-inso/src/cli.js
@@ -1,11 +1,12 @@
 // @flow
 import { ConversionTypeMap, generateConfig } from './commands/generate-config';
-import { getVersion, getAllOptions, logErrorExit1, exit, isDevelopment } from './util';
+import { getVersion, logErrorExit1, exit, isDevelopment } from './util';
 import { runInsomniaTests, TestReporterEnum } from './commands/run-tests';
 import { lintSpecification } from './commands/lint-specification';
 import { exportSpecification } from './commands/export-specification';
 import { parseArgsStringToArgv } from 'string-argv';
 import commander from 'commander';
+import loadOptions from './load-options';
 
 type CreateCommandType = (command?: string, options?: Object) => Object;
 
@@ -26,7 +27,7 @@ function makeGenerateCommand(createCommand: CreateCommandType) {
     )
     .option('-o, --output <path>', 'save the generated config to a file')
     .action((identifier, cmd) =>
-      exit(generateConfig(identifier, getAllOptions(cmd, { type: defaultType }))),
+      exit(generateConfig(identifier, loadOptions(cmd, { type: defaultType }))),
     );
 
   return generate;
@@ -52,7 +53,7 @@ function makeTestCommand(createCommand: CreateCommandType) {
     .option('-b, --bail', 'abort ("bail") after first test failure')
     .option('--keepFile', 'do not delete the generated test file')
     .action((identifier, cmd) =>
-      exit(runInsomniaTests(identifier, getAllOptions(cmd, { reporter: defaultReporter }))),
+      exit(runInsomniaTests(identifier, loadOptions(cmd, { reporter: defaultReporter }))),
     );
 
   return run;
@@ -66,7 +67,7 @@ function makeLintCommand(createCommand: CreateCommandType) {
   lint
     .command('spec [identifier]')
     .description('Lint an API Specification')
-    .action((identifier, cmd) => exit(lintSpecification(identifier, getAllOptions(cmd))));
+    .action((identifier, cmd) => exit(lintSpecification(identifier, loadOptions(cmd))));
 
   return lint;
 }
@@ -80,7 +81,7 @@ function makeExportCommand(createCommand: CreateCommandType) {
     .command('spec [identifier]')
     .description('Export an API Specification to a file')
     .option('-o, --output <path>', 'save the generated config to a file')
-    .action((identifier, cmd) => exit(exportSpecification(identifier, getAllOptions(cmd))));
+    .action((identifier, cmd) => exit(exportSpecification(identifier, loadOptions(cmd))));
 
   return exportCmd;
 }
@@ -93,7 +94,7 @@ function addScriptCommand(originalCommand: Object) {
     .allowUnknownOption()
     .action((scriptName, cmd) => {
       // Load scripts
-      const options = getAllOptions(cmd);
+      const options = loadOptions(cmd);
 
       // Ignore the first arg because that will be scriptName, get the rest
       const passThroughArgs = cmd.args.slice(1);
@@ -129,9 +130,9 @@ function makeDebugCommand(createCommand: CreateCommandType): Object {
     .description('Print all options parsed')
     .allowUnknownOption()
     .action(cmd => {
-      const { __configFile, ...options } = getAllOptions(cmd);
+      const { __configFile, ...options } = loadOptions(cmd);
 
-      console.log(`getAllOptions resolves to:\n${JSON.stringify(options, null, 2)}`);
+      console.log(`loadOptions resolves to:\n${JSON.stringify(options, null, 2)}`);
       console.log(`cmd.args resolves to:\n${JSON.stringify(cmd.args)}`);
 
       return exit(
@@ -143,7 +144,7 @@ function makeDebugCommand(createCommand: CreateCommandType): Object {
     .command('config-file')
     .description('Load the config file')
     .action(cmd => {
-      const options = getAllOptions(cmd);
+      const options = loadOptions(cmd);
 
       console.log(`Loaded config file:\n${JSON.stringify(options.__configFile || {}, null, 2)}`);
       return exit(
